User profile deletion admin log out bug fix.
When admin deleted a profile admin was logged out of the session. It was suppose to only destroy all the sessions of the user whose account is deleted, instead of the current user.

+++ Rails/ROR_Assignment/app/controllers/users_controller.rb
@@ -33,7 +33,9 @@ class UsersController < ApplicationController
 
     def destroy
         @user = User.find(params[:id])
-        session.delete(:user_id)
+        if current_user.nil? == false and current_user.id == @user.id
+            session.delete(:user_id)
+        end
         @user.destroy
         redirect_to users_path
     end
